==> 0-setup_project.Rmd <==
# Simulate dengue model - Setup Project
# ======================================

This Rmarkdown script is used to set-up a dengue modelling project. 

```{r Setup}
# Restart R and set working directory to source file; i.e. rm(list=ls()) 

# Packages ----------------------------------------------------------------
require(tidyverse, quietly = TRUE)
require(readr)
require(lubridate)
require(R.utils)

# Set-up directories ------------------------------------------------------
basePath <- getwd()

outputsPath <- file.path(basePath, "outputs")

# Load model option and function files ------------------------------------
source("code/functions.R")

# Project directory
# project_directory <- file.path(basePath, "projects")
# project_directory

```

## User inputs

The following chunk is where the user enters model simulation details, 
scenario specifications, and outputs 

```{r User inputs and project set-up}

# Model project name
project <- "base"  # options: "randomYears", "testing_caldwell", 
                          #"testing_base", "explore_rainfall"

# startYear <- 1995 # Start Jan 1 this year
# endYear <- 2014 # End Dec 31 this year
# 
# # Simulation time-step in days
# dt <- 1/24
# 
# # Model options
# #### need to be specified ###
model <- "base" # options: "base", "caldwell", "lee"
modelType <- "difference" # options: "ode", "difference", "stochastic"
                          # Note: difference = deterministic,
                          # stochastic = difference with randomness
modelSpecs <- c(model, modelType)

location <- "Dhaka"
# 
# if (startYear == 2030 | startYear == 2080) {
#         gcmOptions <- c("GFDL-ESM2M","HadGEM2-ES","IPSL-CM5A-LR","MIROC5")
# } else if (startYear == 1995) {
#         gcmOptions <- c("Observed","GFDL-ESM2M","HadGEM2-ES","IPSL-CM5A-LR","MIROC5")
# }
# # gcmOptions <- c("Observed","GFDL-ESM2M","HadGEM2-ES","IPSL-CM5A-LR","MIROC5")
# # 
# rcpOptions <- c("Observed","RCP 4.5", "RCP 8.5")
# 
# # get the projected birth and death rate
# projDemRate <- read.csv(file.path(getwd(), "data", "bd_projectedBRDR.csv")) %>%
#   filter(MidPeriod > startYear & MidPeriod < endYear)
# 
# birthRate <- projDemRate$CBR
# 
# deathRate <- projDemRate$CDR
# 
# # Simulation parameters
# sims <- 1000 # Number of simulations

# Set-up model project based on the user inputs above ---------------------

pg <- list()
# pg$sims <- sims
pg$project <- project
# pg$startYear <- startYear
# pg$endYear <- endYear
# pg$dt <- dt 
pg$modelSpecs <- modelSpecs
pg$location <- location
# pg$gcms <- gcmOptions
# pg$rcps <- rcpOptions
# pg$sims <- sims
# pg$birthRate <- birthRate
# pg$deathRate <- deathRate
#   
# # Various time-step options which can be useful. Main one is pts and npts.
# start_date <- as_date(paste0(startYear, "-01-01"))
# end_date <- as_date(paste0(endYear, "-12-31"))
# days <- as_date(start_date:end_date)
# ndays <- length(days)
# 
# pts <- seq(1, ndays, by = pg$dt)
# npts <- length(pts)
# 
# pts_times <- as_datetime(start_date + pts -1)
# 
# pg$start_date <- start_date
# pg$end_date <- end_date
# pg$days <- days
# pg$ndays <- ndays
# pg$pts <- pts
# pg$npts <- npts
# pg$pts_times <- pts_times

# Set-up folders for outputs
projectFolder <- file.path(outputsPath, project)
dir.create(projectFolder, showWarnings = FALSE)
dir.create(file.path(projectFolder, "results"), showWarnings = FALSE)

```

```{r set seed}
# Set seed (this will be saved in pg which can be re-used by setting 
# re-sample to FALSE - needed for sampling and stochastic simulations)
reseed <- TRUE 
if (reseed) {
        sim_seed <- runif(1, 1, 1e6)
} else {
        sim_seed <- pg$sim_seed
}
set.seed(sim_seed)
pg$sim_seed <- sim_seed

```


```{r Copy parameter files}
# Copy template parameter files to project directory. These must 
# then be filled in by the user creating default parameters for the
# project

file.copy(file.path(basePath, "templates", "parameters.csv"), 
  file.path(projectFolder, "parameters.csv"), overwrite = TRUE)

if (model == "caldwell") {
        file.copy(file.path(basePath, "templates", "caldwell_parameters.csv"), 
                  file.path(projectFolder, "caldwell_parameters.csv"),
                  overwrite = TRUE)
}
if (model == "base") {
        file.copy(file.path(basePath, "templates", "base_parameters.csv"), 
                  file.path(projectFolder, "base_parameters.csv"),
                  overwrite = TRUE)
}

```

```{r Save project specs}

save(pg, file = file.path(projectFolder, "Project_specifications.Rda"))

```
