==> 5-observed_analysis.Rmd <==
# Simulate dengue model - Analysis of model output for 1995-2014
# ==============================================================

This Rmarkdown script is used to analyse model output for the period 1995-2014.

```{r sim summary}

require(tidyverse, quietly = TRUE)
require(lubridate)
require(cowplot)

source("code/PlotOptions.R")

# folder in which simulations are saved

basePath <- getwd()
outputsPath <- file.path(basePath, "outputs")
project <- "base"
projectFolder <- file.path(outputsPath, project)

load(file.path(projectFolder, "Model_inputs_calibrated.Rda"))

saveSims <- FALSE

period <- "1995_2014" # "2080_2099", "1995_2014"
RCP <-  "Observed"# "RCP 4.5", "RCP 8.5"
baseline_sero <- "low_baseline_seroprev" # "high_baseline_seroprev"


if (period == "1995_2014") {
        simFolder <- file.path(projectFolder, "savedsims", paste0(period))
} else {
        simFolder <- file.path(projectFolder, "savedsims", paste0(period), paste0(RCP)
                               #, paste0(baseline_sero)
                               )
}

Dat <- list()

for (ii in 1:1000) {
  print(ii)
        for (jj in pg$gcms) {
                tempScenario <- ifelse(jj == "Observed", "Observed", paste0(RCP))
                for(kk in tempScenario) {
                        load(file.path(simFolder, paste0(ii, "_", jj, "_", kk, ".Rda")))
                        name <- paste(ii,jj,kk)
                        Dat[[name]] <- adjustedDat
                }
        }
}

# Convert lists into date frames - will get a warning but no issue 
datDf <- unnest(enframe(map(Dat, "dat"))) %>%
        rename("model" = name) %>%
        select("model", "newInfections", "newInfections1", "newInfections2",  
               "date", "month", "year", "Nh", "Nv", "seroPrev", "days")

annInfectionsDf <- unnest(enframe(map(Dat, "annInfections"))) %>%
  rename("model" = name)

monthInfectionsDf <- unnest(enframe(map(Dat, "monthInfections"))) %>%
  rename("model" = name)

seroprevDf <- unnest(enframe(map(Dat, "seroprev"))) %>%
  rename("model" = name)

# save sim summary
if (saveSims){
  save(datDf, annInfectionsDf, monthInfectionsDf, seroprevDf, 
       file = file.path(simFolder, "sim_summary.Rda"))
} else {
  load(file.path(simFolder,"sim_summary.Rda"))
}

# Create Model and Set variable from combined model variable
datDf[c('Sim', 'GCM', 'RCP')] <- str_split_fixed(datDf$model, ' ', 3)

annInfectionsDf[c('Sim', 'GCM', 'RCP')] <- str_split_fixed(annInfectionsDf$model, ' ', 3)

monthInfectionsDf[c('Sim', 'GCM', 'RCP')] <- str_split_fixed(monthInfectionsDf$model, ' ', 3)

seroprevDf[c('Sim', 'GCM', 'RCP')] <- str_split_fixed(seroprevDf$model, ' ', 3)

# seroprevHist <- ggplot(seroprevDf, aes(x=value))+
#   geom_histogram(bins = 100)+
#   xlab("Seroprevalence (%)") + ylab("Frequency") +
#   facet_wrap(~GCM)+
#   PlotOptions()
# 
# seroprevHist

# Plot specifications
yearPoints <- c(0.5,4.5,8.5,12.5,16.5,19.5)

if (period == "2080_2099") {
        yearLabels <- c("2080", "2084", "2088", "2092", "2096", "2099")
} else if (period == "2030_2049"){
        yearLabels <- c("2030", "2034", "2038", "2042", "2046", "2049")   
} else {
        yearLabels <- c("1995", "1999", "2003", "2007", "2011", "2014") 
}

ptsPlots <- yearPoints*365

#### Reported dengue data
monthlyDiagnoses <- read.csv(file.path(getwd(), "data", "Monthly_dengue_cases_2000-2019.csv")) %>%
  mutate(time = Year + Month/12)

monthlyDiagnosesPlot <- ggplot(monthlyDiagnoses, aes(x = Year+Month/12)) +
        geom_line(aes(y = Den_cases), colour="blue") +
        xlab("") + ylab("Dengue cases") + 
        
        coord_cartesian(xlim = c(1995, 2014), 
                        ylim = quantile(monthlyDiagnoses$Den_cases, c(0.1, 0.983))) +
        scale_x_continuous(breaks = c(1995.5, 1999.5, 2003.5, 2007.5, 2011.5, 2014.5), 
                           labels = yearLabels)+
        theme(text = element_text(face = "bold",size=18,colour="black"),
              panel.background = element_blank(),
              axis.line = element_line(colour = "black"))

monthlyDiagnosesPlot

seasonalDiagnosis <- ggplot(monthlyDiagnoses %>%
                                    group_by(Month) %>%
                                    summarise(lwr_IQR = quantile(Den_cases, 0.25, na.rm = TRUE),
                                              medInf = median(Den_cases),
                                              upr_IQR = quantile(Den_cases, 0.75, na.rm = TRUE))) +
        geom_line(aes(Month, medInf, color = "Median"))+
        geom_point(aes(Month, medInf))+
        geom_ribbon(aes(x = Month, ymin = lwr_IQR, ymax = upr_IQR,  fill = "iqr",  alpha = "iqr"))+
        xlab("") + ylab("Monthly diagnosis")+
        scale_x_continuous(breaks = round(seq(min(monthlyDiagnoses$Month),
                                              max(monthlyDiagnoses$Month), by = 1),1),
                           labels = c("J","F","M","A","M","J","J","A","S","O","N","D"))+
        scale_colour_manual(NULL, values = "blue",
                            limits = "Median") +
        scale_alpha_manual(NULL, values = 0.4,
                           limits = "iqr",
                           labels = "Interquartile range") +
        scale_fill_manual(NULL, values = "black",
                          limits = "iqr",
                          labels = "Interquartile range") +
        theme(text = element_text(face = "bold",size=18,colour="black"),
              panel.background = element_blank(),
              axis.line = element_line(colour = "black"),
              legend.text=element_text(face = "bold",size=18),
              legend.position = "top")

seasonalDiagnosis

observed_data <- plot_grid(monthlyDiagnosesPlot, seasonalDiagnosis, 
                              labels = c('A', 'B'),
                              hjust = -0.8, vjust = 2, label_size=14)

observed_data

# ggsave(file.path(projectFolder, "Paper figure", "observed_data.png"),
#        observed_data , width = 35, height = 15, units = "cm")
#####

monthlyInfectionsPlot <- ggplot(monthInfectionsDf %>% 
                          group_by(year, month, GCM) %>% 
                          summarise(lwr_IQR = quantile(infections, 0.25, na.rm = TRUE),
                                    medNewInf = median(infections),
                                    upr_IQR = quantile(infections, 0.75, na.rm = TRUE))) +
        geom_line(mapping = aes(x=year+month/12, y=medNewInf, group = GCM, color = "Median"))+
        geom_ribbon(mapping = aes(x = year+month/12, ymin = lwr_IQR, ymax = upr_IQR,  fill = "iqr",  alpha = "iqr"))+
        xlab("") + ylab("Dengue infections") +
        scale_x_continuous(breaks = c(1995.5, 1999.5, 2003.5, 2007.5, 2011.5, 2014.5), 
                           labels = yearLabels) +
        scale_colour_manual(NULL, values = "blue",
                            limits = "Median") +
        scale_alpha_manual(NULL, values = 0.4,
                           limits = "iqr",
                           labels = "Interquartile Range") +
        scale_fill_manual(NULL, values = "black",
                          limits = "iqr",
                          labels = "Interquartile Range") +
        theme(text = element_text(face = "bold",size=18,colour="black"),
              panel.background = element_blank(),
              axis.line = element_line(colour = "black"),
              legend.text=element_text(face = "bold",size=18),
              legend.position = c(0.6, 0.7))  
monthlyInfectionsPlot

dailyInfectionsPlot <- ggplot(datDf %>%
                                      group_by(days, GCM) %>%
                                      summarise(lwr_IQR = quantile(newInfections, 0.25, na.rm = TRUE),
                                                medNewInf = median(newInfections),
                                                upr_IQR = quantile(newInfections, 0.75, na.rm = TRUE))) +
        geom_line(aes(x=days,y=medNewInf, group = GCM, color = "Median"))+
        geom_ribbon(aes(x = days, ymin = lwr_IQR, ymax = upr_IQR,  fill = "iqr",  alpha = "iqr"))+
        xlab("") + ylab("Daily dengue infections") +
        scale_x_continuous(breaks = ptsPlots, 
                           labels = yearLabels) +
        scale_colour_manual(NULL, values = "blue",
                            limits = "Median") +
        scale_alpha_manual(NULL, values = 0.4,
                           limits = "iqr",
                           labels = "Interquartile Range") +
        scale_fill_manual(NULL, values = "black",
                          limits = "iqr",
                          labels = "Interquartile Range") +
        coord_cartesian(xlim = c(1, ptsPlots[6]))+
        theme(text = element_text(face = "bold",size=14,colour="black"),
              panel.background = element_blank(),
              axis.line = element_line(colour = "black"),
              legend.position = c(0.6, 0.7))

dailyInfectionsPlot

seroPrevPlot <- ggplot(datDf %>%
                               group_by(days, GCM) %>%
                               summarise(lwr_IQR = quantile(seroPrev, 0.25, na.rm = TRUE),
                                         medseroPrev = median(seroPrev),
                                         upr_IQR = quantile(seroPrev, 0.975, na.rm = TRUE))) +
        geom_line(aes(x=days,y=medseroPrev, group = GCM, color = "Median Seroprevalence"))+
        geom_ribbon(aes(x = days, ymin = lwr_IQR, ymax = upr_IQR, fill = "iqr",  alpha = "iqr"))+
        geom_segment(aes(x = ptsPlots[6], y = 36, xend = ptsPlots[6], yend = 85), color = "red")+
        xlab("") + ylab("Seroprevalence (%)")+
        scale_x_continuous(breaks = ptsPlots, 
                           labels = yearLabels) +
        scale_colour_manual(NULL, values = "blue",
                            limits = "Median Seroprevalence") +
        scale_alpha_manual(NULL, values = 0.4,
                           limits = "iqr",
                           labels = "Interquartile Range") +
        scale_fill_manual(NULL, values = "black",
                          limits = "iqr",
                          labels = "Interquartile Range") +
        coord_cartesian(xlim = c(1, ptsPlots[6]))+
        theme(text = element_text(face = "bold",size=18,colour="black"),
              panel.background = element_blank(),
              axis.line = element_line(colour = "black"),
              legend.position = "none")

seroPrevPlot

annInfectionPlot <- ggplot(annInfectionsDf, aes(x = GCM, y = infections)) +
        geom_boxplot(outlier.shape = NA)+
        scale_y_continuous(limits = quantile(annInfectionsDf$infections, c(0.1, 0.9)))+
        xlab("") + ylab("Total annual infections") +
        theme(text = element_text(face = "bold",size=18,colour="black"),
              panel.background = element_blank(),
              axis.line = element_line(colour = "black"),
              axis.title.x=element_blank(),
              axis.text.x=element_blank(),
              axis.ticks.x=element_blank())

annInfectionPlot

###
#boxplot statistics
bp <- boxplot(annual_peak$x)
outliers <- bp$out
num_outliers <- length(outliers)
max_value <- range(annual_peak$x, na.rm = TRUE)[2]
cat("Number of outliers:", num_outliers, "\n")
cat("Maximum value:", max_value)

###
annual_peak <- aggregate(datDf$newInfections, list(datDf$year,datDf$Sim, datDf$GCM), FUN=max)

PeakInfectionPlot <- ggplot(annual_peak, aes(x = Group.3, y = x)) +
        geom_boxplot(outlier.shape = NA)+
        scale_y_continuous(limits = quantile(annual_peak$x, c(0.1, 0.9)))+
        xlab("") + ylab("Annual peak size") +
        theme(text = element_text(face = "bold",size=18,colour="black"),
              panel.background = element_blank(),
              axis.line = element_line(colour = "black"),
              axis.title.x=element_blank(),
              axis.text.x=element_blank(),
              axis.ticks.x=element_blank())

PeakInfectionPlot
        
seasonalPlot <- ggplot(monthInfectionsDf %>%
                               group_by(month, GCM) %>%
                               summarise(lwr_IQR = quantile(infections, 0.25, na.rm = TRUE),
                                         medInf = median(infections),
                                         upr_IQR = quantile(infections, 0.75, na.rm = TRUE))) +
        geom_line(aes(month, medInf, group = GCM, color = "Median monthly infections"))+
        geom_point(aes(month, medInf))+
        geom_ribbon(aes(x = month, ymin = lwr_IQR, ymax = upr_IQR,  fill = "iqr",  alpha = "iqr"))+
        xlab("") + ylab("Monthly infections")+
        scale_x_continuous(breaks = round(seq(min(monthInfectionsDf$month),
                                              max(monthInfectionsDf$month), by = 1),1),
                           labels = c("J","F","M","A","M","J","J","A","S","O","N","D"))+
        scale_colour_manual(NULL, values = "blue",
                            limits = "Median monthly infections") +
        scale_alpha_manual(NULL, values = 0.4,
                           limits = "iqr",
                           labels = "IQR") +
        scale_fill_manual(NULL, values = "black",
                          limits = "iqr",
                          labels = "IQR") +
        theme(text = element_text(face = "bold",size=18,colour="black"),
              panel.background = element_blank(),
              axis.line = element_line(colour = "black"),
              legend.position = "none")

seasonalPlot

p1 <- plot_grid(monthlyInfectionsPlot, seroPrevPlot, labels = c('A', 'B'), 
                rel_widths = c(1, 1/2), hjust = -7, label_size=18, align = "h" )
p2 <- plot_grid(annInfectionPlot, PeakInfectionPlot, seasonalPlot,  
                labels = c('C', 'D', 'E'), hjust = -7, label_size=18, nrow = 1)

obs_uncertainity <- plot_grid(p1, p2, align = "v", ncol = 1)

obs_uncertainity

## important chapter 2 result figure
# ggsave(file.path(projectFolder, "Paper figure", "obs_uncertainity.png"),
#        obs_uncertainity , width = 35, height = 20, units = "cm")

```
